Link 100%

+++ over-hal/scan_address.h
@@ -4,4 +4,4 @@
 #define scan_address_ip(s,a) scan_address (s, '.', a, 10)
 #define scan_address_mac(s,a) scan_address(s, ':', a, 16)
 
-void scan_address(char * StringToScan, char CharSep, unsigned char *Array, int Base);
+void scan_address(char * StringToScan, char CharSep, unsigned char *Array, unsigned short Base);
